move sort function in commonb.js utility script
